refactor: consolidate play queuing into abstract component
