+++ src/backend/scrobblers/AbstractScrobbleClient.ts
@@ -2,7 +2,6 @@ import { childLogger, type Logger, type LogLevel } from "@foxxmd/logging";
 import dayjs, { type Dayjs } from "dayjs";
 import type {Duration} from "dayjs/plugin/duration.js";
 import type EventEmitter from "events";
-import { nanoid } from "nanoid";
 import type { MarkOptional } from "ts-essentials";
 import {
     type NowPlayingUpdateThreshold,
@@ -14,11 +13,10 @@ import {
     QUEUE_STATUS_COMPLETED,
     SOURCE_SOT,
     QUEUE_STATUS_FAILED,
-    isPlayObject,
     DEAD_LETTER_RETRIES_DEFAULT,
     PARSED_FROM
 } from "../../core/Atomic.ts";
-import { buildTrackString, capitalize, truncateStringToLength } from "../../core/StringUtils.ts";
+import { buildTrackString, truncateStringToLength } from "../../core/StringUtils.ts";
 import AbstractComponent from "../common/AbstractComponent.ts";
 import { hasUpstreamError } from "../common/errors/UpstreamError.ts";
 import {
@@ -55,26 +53,23 @@ import { todayAwareFormat } from "../../core/TimeUtils.ts";
 import { AsyncTask, SimpleIntervalJob, ToadScheduler } from "toad-scheduler";
 import { getRoot } from "../ioc.ts";
 import { staggerMapper, type StaggerOptions } from "../utils/AsyncUtils.ts";
-import pMap, { pMapIterable } from "p-map";
+import pMap from "p-map";
 import { existingScrobble, type ExistingScrobbleOpts } from "../utils/PlayComparisonUtils.ts";
 import { statefulInvariantTransform } from "../../core/PlayUtils.ts";
-import { normalizeStr } from "../utils/StringUtils.ts";
 import type { Counter, Gauge } from 'prom-client';
 import { generateLoggableAbortReason, ScrobbleSubmitError, SimpleError, StageChangeError } from "../common/errors/MSErrors.ts";
 import { serializeError} from 'serialize-error';
 import { DEFAULT_NEW_PADDING, groupPlaysToTimeRanges } from "../utils/ListenFetchUtils.ts";
 import { spawn, isAbortError, delay, waitForEvent } from 'abort-controller-x';
-import { DrizzlePlayRepository, playToRepositoryCreatePlayOpts, type QueryPlaysOpts, type WithPlayRelation } from "../common/database/drizzle/repositories/PlayRepository.ts";
-import type {PlayEventNew, PlayEventSelect, PlaySelect, PlaySelectWithQueueStates, PlayWith, QueueStateSelect} from "../common/database/drizzle/drizzleTypes.ts";
+import { type QueryPlaysOpts, type WithPlayRelation } from "../common/database/drizzle/repositories/PlayRepository.ts";
+import type {PlayEventNew, PlayEventSelect, PlaySelectWithQueueStates, PlayWith, QueueStateSelect} from "../common/database/drizzle/drizzleTypes.ts";
 import { asPlay } from "../../core/PlayMarshalUtils.ts";
-import { DrizzleQueueRepository } from "../common/database/drizzle/repositories/QueueRepository.ts";
 import { GenericRepository } from "../common/database/drizzle/repositories/BaseRepository.ts";
 import assert from "node:assert";
 import { COMPONENT_STATE, type ComponentClientApiJson, type PlayApiCommonDetailed, type QueueStateApi } from "../../core/Api.ts";
 import type {ComponentState} from "react";
-import { DrizzlePlayEventsRepository } from "../common/database/drizzle/repositories/PlayEventsRepository.ts";
 import { PLAY_EVENT_TYPE, type PlayEvent } from "../../core/PlayEvent.ts";
-import { dupeCheckToPlayEvent, entityIsPlayEntity, queueCompletionStateToPlayEvent, queueStateToPlayEvent, scrobbleToPlayEvent, stateChangeToPlayEvent, transformToPlayEvent } from "../common/database/drizzle/entityUtils.ts";
+import { dupeCheckToPlayEvent, queueCompletionStateToPlayEvent, queueStateToPlayEvent, scrobbleToPlayEvent, stateChangeToPlayEvent, transformToPlayEvent } from "../common/database/drizzle/entityUtils.ts";
 import type { PlayProcessingResult } from "../common/infrastructure/PlayProcessing.ts";
 import { PlayProcessingError } from "../common/errors/PlayProcessingError.ts";
 
@@ -122,7 +117,7 @@ export default abstract class AbstractScrobbleClient extends AbstractComponent i
     scrobbleRetries: number =  0;
     scrobbling: boolean = false;
     deadQueueProcessing: boolean = false;
-    queuedLength: number = 0;
+    //queuedLength: number = 0;
     deadLetterLength: number = 0;
     deadLetterQueued: number  = 0;
 
@@ -149,7 +144,7 @@ export default abstract class AbstractScrobbleClient extends AbstractComponent i
     notifier: Notifiers;
 
     protected scrobbledCounter: Counter;
-    protected queuedGauge: Gauge;
+    //protected queuedGauge: Gauge;
     protected deadLetterGauge: Gauge;
     protected problemGauge: Gauge;
 
@@ -451,23 +446,6 @@ export default abstract class AbstractScrobbleClient extends AbstractComponent i
             }
             this.staggerMappers.existing = staggerMapper<PlayObject, PlayObject>({initialInterval: Math.max(...eInits), maxRandomStagger: Math.max(...eMaxStagger), concurrency: 3});
         }
-    }
-
-    protected getIdentifier() {
-        return `${capitalize(this.type)} - ${this.name}`
-    }
-    protected getMachineId() {
-        return `${this.type}-${this.name}`;
-    }
-    public getSafeExternalName() {
-        return normalizeStr(this.name, {keepSingleWhitespace: false});
-    }
-    public getSafeExternalId() {
-        return `${this.type}-${normalizeStr(this.name, {keepSingleWhitespace: false})}`;
-    }
-
-    protected getPrometheusLabels() {
-        return {name: this.getSafeExternalName(), type: this.type};
     }
 
     public getRunningState(): ComponentState {
@@ -1167,7 +1145,7 @@ export default abstract class AbstractScrobbleClient extends AbstractComponent i
             let more = true;
             let offset = 0;
             while(more) {
-                await this.queueScrobble(processable.data, {reason});
+                await this.queuePlay(processable.data, {reason});
                 more = processable.data.length === processable.meta.limit;
                 if(more) {
                     offset += processable.meta.limit;
@@ -1438,109 +1416,57 @@ export default abstract class AbstractScrobbleClient extends AbstractComponent i
         }
     }
 
-    queueScrobble = async (data: (PlayObject | PlayObject[]) | (PlaySelectWithQueueStates | PlaySelectWithQueueStates[]), context?: QueueContext & {isRetry?: boolean}) => {
-        const createdQueuedPlays: PlaySelect[] = [];
+    async findPreQueueExistingPlay(queueablePlay: PlayObject, context?: QueueContext & {isRetry?: boolean}) {
+        /**
+        * WHEN RUN IN A CLIENT COMPONENT
+        * 
+        * we *always* do checkExisting because by the time we are queueing plays
+        * * they should have already passed this check in a Source
+        * * or user has manually/forced a Play to be discovered on a Source
+        * * or we are running a migration job (future functionality)
+        * 
+        * and clients will never have the same level of noise/repeat polling as a source so its more useful to store and log all dupes
+        */
 
-        const dataArray = Array.isArray(data) ? data : [data];
-
-        if (dataArray.every(x => entityIsPlayEntity(x))) {
-            for (const playSelect of dataArray) {
-                let queue = playSelect.queueStates.find(x => x.queueName === INGRESS_QUEUE);
-                if (queue === undefined) {
-                    queue = await this.queueRepo.create({ componentId: this.dbComponent.id, playId: playSelect.id, queueName: INGRESS_QUEUE, context }) as QueueStateSelect;
-                } else {
-                    this.queueRepo.updateById(queue.id, { queueStatus: 'queued', context, error: undefined });
-                }
-                const events = await this.playEventsRepo.createMany([
-                    { playId: playSelect.id, ...stateChangeToPlayEvent({ state: 'queued' }) },
-                    { playId: playSelect.id, ...queueStateToPlayEvent({ ...queue, queueStatus: 'queued', error: undefined, context: context ?? queue.context }) }
-                ]) as PlayEventSelect[];
-                playSelect.state = 'queued';
-                await this.playRepo.updateById(playSelect.id, {state: 'queued'});
-                if (`events` in playSelect) {
-                    (playSelect as PlayWith<'events'>).events = (playSelect as PlayWith<'events'>).events.concat(events);
-                } else {
-                    (playSelect as unknown as PlayWith<'events'>).events = events;
-                }
-                this.emitPlayUpdate({ ...playSelect } as unknown as PlayApiCommonDetailed);
-                this.emitEvent(queue.retries > 0 ? 'deadQueued' : 'playQueued', {queuedPlay: playSelect});
-                createdQueuedPlays.push(playSelect);
-            }
-        } else if (dataArray.every(x => isPlayObject(x))) {
-            const monitoring = this.getMonitoringStatus();
-            const {
-                transform = true,
-            } = context || {};
-            const playDatas = dataArray.map(x => ({...x, meta: {...x.meta, wasMonitored: monitoring.monitoring, seenAt: dayjs()}}));
-
-            for await(const play of pMapIterable(playDatas, this.staggerMappers.preCompare(async x => transform === false ? await noopTransform(x) : await this.transformPlay(x, TRANSFORM_HOOK.preCompare)), {concurrency: 3})) {
-                try {
-                    // cheap check, looks for play data (non-meta) hash, playdate, and optionally mbid recording
-                    const cheapExisting = await this.playRepo.checkExisting(play, { queueName: INGRESS_QUEUE });
-                    if (cheapExisting !== undefined) {
-                        const qs = cheapExisting.queueStates.find(x => x.queueName === INGRESS_QUEUE);
-                        this.logger.trace(`Not adding to queue because it is already in the queue, discovered via hash/mbid, last queued at ${todayAwareFormat(qs.createdAt)}`);
-                        continue;
-                    }
-                    // then chunked queued plays
-                    let offset = 0;
-                    let inQueue = false;
-                    while (true) {
-                        const { data, meta } = await this.playRepo.getQueued(INGRESS_QUEUE, { offset, retries: 0 });
-                        const existingQueued = await this.existingPlay(play, data.map(x => asPlay(x.play)), false);
-                        // want to be very confident of this
-                        if (existingQueued.match && existingQueued.score > 0.99) {
-                            this.logger.trace(`Not adding to queue because it is already in the queue\n${existingQueued.summary}`);
-                            inQueue = true;
-                            break;
-                        }
-                        if (data.length < meta.limit) {
-                            break;
-                        }
-                        offset += meta.limit;
-                    }
-
-                    if (inQueue) {
-                        continue;
-                    }
-                } catch (e) {
-                    this.logger.warn(new SimpleError('Failed to check queued scrobble for existing before adding, will continue with adding anyway', { cause: e }));
-                }
-                // not in queue or existing queued check failed for some reason and we don't want to lose scrobble
-                const {
-                    data,
-                    meta
-                } = play
-                const createPlayData = playToRepositoryCreatePlayOpts({
-                    play: {
-                        data,
-                        meta
-                    },
-                    componentId: this.dbComponent.id, 
-                    state: 'queued',
-                    parentId: play.id
-                });
-
-                const playRow = await this.playRepo.createPlays([createPlayData]);
-                const queueState = await this.queueRepo.create({componentId: this.dbComponent.id, playId: playRow[0].id, queueName: INGRESS_QUEUE, context}) as QueueStateSelect;
-                const createdEvents = await this.playEventsRepo.createMany([
-                    {playId: playRow[0].id, ...stateChangeToPlayEvent({state: 'queued'}), createdAt: playRow[0].seenAt.add(1,'ms')},
-                    {playId: playRow[0].id, ...queueStateToPlayEvent(queueState), createdAt: queueState.createdAt}
-                ]);
-                createdQueuedPlays.push(playRow[0]);
-                this.logger.debug(`Added ${buildTrackString(play)} to the queue`);
-                this.setStatus(`Added Play from parent ${play.uid} to queue`);
-
-                const queuedPlay = {id: nanoid(), source: meta.source, play: play}
-                this.emitEvent('playQueued', {queuedPlay: queuedPlay});
-                this.emitPlayInsert({...playRow[0], queueStates: [queueState], events: createdEvents} as unknown as PlayApiCommonDetailed);
-                this.queuedLength += 1;
-                this.queuedGauge.labels(this.getPrometheusLabels()).inc();
-            }
-        } else {
-            throw new Error('Data passed to queuePlay must be either all be PlayObject or all PlaySelect objects');
+        // we should be adding Plays to the queue without any transforms
+        // so run on "raw" play input
+        // if we have seen a play with close temporality with the exact input hash then skip it entirely
+        const cheapInputExisting = await this.playRepo.checkExisting(queueablePlay, {
+                inputHash: queueablePlay,
+                // existing should have been created *before* this play
+                seenAt: {
+                    type: 'lt',
+                    date: dayjs(),
+                },
+                // if this play is passed by a source
+                // then the playobject has the id of the source (parent) play
+                // and we can additionally check for frozen source play hash to bypass differences
+                // caused by lifecycle transforms
+                parentId: queueablePlay.id,
+            });
+        if (cheapInputExisting !== undefined) {
+            this.logger.trace(`Not adding ${buildTrackString(queueablePlay)} to queue because it already exists in db as Play ${cheapInputExisting.uid}`);
+            return cheapInputExisting;
         }
-        return createdQueuedPlays;
+
+        // may not be needed but if not existing cheap then
+        // we run a second check against only queued plays
+        // with the more comprehensive scoring dupe check
+        let offset = 0;
+        while (true) {
+            const { data, meta } = await this.playRepo.getQueued(INGRESS_QUEUE, { offset, retries: 0 });
+            const existingQueued = await this.existingPlay(queueablePlay, data.map(x => asPlay(x.play)), false);
+            // want to be very confident of this
+            if (existingQueued.match && existingQueued.score > 0.99) {
+                this.logger.trace(`Not adding to queue because it is already in the queue\n${existingQueued.summary}`);
+                return await this.playRepo.findByIdWith<'queueStates' | 'parent'>(existingQueued.closestMatchedPlay.id, ['queues','parent']);
+            }
+            if (data.length < meta.limit) {
+                break;
+            }
+            offset += meta.limit;
+        }
+        return;
     }
 
     queuePlayingNow = async (data: SourcePlayerObj, source: SourceIdentifier) => {
