+++ src/backend/scrobblers/ScrobbleClients.ts
@@ -453,7 +453,7 @@ export default class ScrobbleClients {
                 }
             }
             for (const playObj of playObjs) {
-                await client.queueScrobble(clone(playObj));
+                await client.queuePlay(clone(playObj));
             }
         }
         if(excluded.length > 0) {
